% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcApprovals.R
\docType{data}
\name{nhmrcApprovals}
\alias{nhmrcApprovals}
\title{NHMRC Approvals.}
\format{Data frame with columns
\describe{
\item{grant_id}{Grant ID}
\item{portfolio}{Portfolio (Health)}
\item{agency}{Agency (NHMRC)}
\item{program_sub_title}{Program sub-title}
\item{recipient_rao}{Recipient RAO}
\item{purpose}{Purpose}
\item{recipient_institution}{Recipient institution}
\item{value}{Value awarded}
\item{approval_date}{Approval date}
\item{grant_term_months}{Grant term months}
\item{grant_funding_location}{Grant funding location}
\item{postcode}{Postcode}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/grant_approvals_15-2-17.xlsx
}
\usage{
nhmrcApprovals
}
\description{
Compliance reporting of grant outcomes.
}
\examples{
  nhmrcApprovals \%>\% glimpse()
}
\keyword{datasets}
