% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcBRA.R
\docType{data}
\name{nhmrcBRA}
\alias{nhmrcBRA}
\title{NHMRC Expenditure by Broad Research Area.}
\format{Data frame with columns
\describe{
\item{calendar_year}{Calendar year}
\item{basic_research_area}{Basic research area (BRA)}
\item{value}{Expenditure for the BRA}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/dataset/2015/2._summary_tables_2000_-_2015.xlsx
}
\usage{
nhmrcBRA
}
\description{
Summary of expenditure by broad research area and year.
}
\examples{
  nhmrcBRA
}
\keyword{datasets}
