% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcExpenditure.R
\docType{data}
\name{nhmrcExpenditure}
\alias{nhmrcExpenditure}
\title{NHMRC Expenditure.}
\format{Data frame with columns
\describe{
\item{calendar_year}{Calendar year}
\item{total_expenditure}{Total expenditure for the calendar year}
\item{active_grants}{Active grants for the calendar year}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/dataset/2016/all_grants_2000-2016.xlsx
}
\usage{
nhmrcExpenditure
}
\description{
Summary of total expenditure and active grants by year.
}
\examples{
  nhmrcExpenditure
}
\keyword{datasets}

