% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcFundingType.R
\docType{data}
\name{nhmrcFundingType}
\alias{nhmrcFundingType}
\title{NHMRC Expenditure by Funding Type.}
\format{Data frame with columns
\describe{
\item{calendar_year}{Calendar year}
\item{support}{Type of funding}
\item{value}{Expenditure for funding type}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/dataset/2015/2._summary_tables_2000_-_2015.xlsx
}
\usage{
nhmrcFundingType
}
\description{
Summary of expenditure by funding type and year.
}
\examples{
  nhmrcFundingType \%>\% glimpse()
}
\keyword{datasets}
