% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcGrantType.R
\docType{data}
\name{nhmrcGrantType}
\alias{nhmrcGrantType}
\title{NHMRC Expenditure by Grant Type.}
\format{Data frame with columns
\describe{
\item{year}{Calendar year}
\item{main_funding_group}{Type of funding}
\item{lower_grant_type}{Grant type}
\item{value}{Expenditure for grant type}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/dataset/2016/all_grants_2000-2016.xlsx
}
\usage{
nhmrcGrantType
}
\description{
Summary of expenditure by grant type and year.
}
\examples{
  nhmrcGrantType
}
\keyword{datasets}
