% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcInstitute.R
\docType{data}
\name{nhmrcInstitute}
\alias{nhmrcInstitute}
\title{NHMRC Expenditure by Administering Institution.}
\format{Data frame with columns
\describe{
\item{year}{Calendar year}
\item{administering_institution}{Administering institution}
\item{value}{Expenditure for grant subtype}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/dataset/2016/all_grants_2000-2016.xlsx
}
\usage{
nhmrcInstitute
}
\description{
Summary of expenditure by institution and year.
}
\examples{
  nhmrcInstitute
}
\keyword{datasets}
