% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcOutcomesGenderBRA.R
\docType{data}
\name{nhmrcOutcomesGenderBRA}
\alias{nhmrcOutcomesGenderBRA}
\title{NHMRC Outcomes Gender BRA.}
\format{Data frame with columns
\describe{
\item{year}{Year}
\item{scheme}{Scheme}
\item{category}{Category}
\item{gender}{Gender}
\item{stage}{Stage (applied or funded)}
\item{value}{Amount awarded}
}}
\source{
2013 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/table_2_women_in_science_140219.pdf

2014 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/table_2_bra_data_scheme_by_gender_141205.pdf

2015 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/webchange_wihs_2015_outcomes_by_gender_table_2.pdf
}
\usage{
nhmrcOutcomesGenderBRA
}
\description{
NHMRC funding outcomes by gender and broad research area for early career and career development fellows.
}
\examples{
  nhmrcOutcomesGenderBRA \%>\% glimpse()
}
\keyword{datasets}
