% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcOutcomesGenderFellows.R
\docType{data}
\name{nhmrcOutcomesGenderFellows}
\alias{nhmrcOutcomesGenderFellows}
\title{NHMRC Outcomes Gender Fellow}
\format{Data frame with columns
\describe{
\item{year}{Year}
\item{scheme}{Scheme}
\item{level}{Level}
\item{gender}{Gender}
\item{stage}{Stage (applied or funded)}
\item{value}{Amount awarded}
}}
\source{
2013 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/table_4_women_in_science_140219.pdf

2014 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/table_4_gender_outcomes_prac_research_cdf_fellows_level_141205.pdf

2015 https://www.nhmrc.gov.au/_files_nhmrc/file/your_health/may_2016_update_table_4_2015_gender_outcomes_for_prac_research_and_cdf_by_level.pdf
}
\usage{
nhmrcOutcomesGenderFellows
}
\description{
NHMRC funding outcomes by gender for Practitioner, Research and Career Development Fellows by Level.
}
\examples{
  nhmrcOutcomesGenderFellows
}
\keyword{datasets}
