% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcOutcomesGenderGrants.R
\docType{data}
\name{nhmrcOutcomesGenderGrants}
\alias{nhmrcOutcomesGenderGrants}
\title{NHMRC Outcomes Gender Grants.}
\format{Data frame with columns
\describe{
\item{year}{Year}
\item{ci}{CI group (CIA, CIB...)}
\item{gender}{Gender}
\item{value}{CI count for each group}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/project_grant_funded_rate_gender_new_investigator_and_teaching_load_140218.xlsx
}
\usage{
nhmrcOutcomesGenderGrants
}
\description{
NHMRC project grants new investigators grant data CI gender 2003 - 2013.
}
\examples{
  nhmrcOutcomesGenderGrants \%>\% glimpse()
}
\keyword{datasets}
