% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcOutcomesGenderPartTime.R
\docType{data}
\name{nhmrcOutcomesGenderPartTime}
\alias{nhmrcOutcomesGenderPartTime}
\title{NHMRC Outcomes Gender Part-time.}
\format{Data frame with columns
\describe{
\item{year}{Year}
\item{scheme}{Scheme}
\item{category}{Category}
\item{gender}{Gender}
\item{stage}{Stage (applied or funded)}
\item{value}{Amount awarded}
}}
\source{
2013 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/table_3_women_in_science_140218.pdf

2014 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/table_3_part_time_fellowship_outcomes_gender_141205.pdf

2015 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/webchange_wihs_2015_outcomes_by_gender_table_3_0.pdf
}
\usage{
nhmrcOutcomesGenderPartTime
}
\description{
NHMRC funding outcomes by gender for part-time fellowships.
}
\examples{
  nhmrcOutcomesGenderPartTime \%>\% glimpse()
}
\keyword{datasets}
