% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcOutcomesGenderPostDoc.R
\docType{data}
\name{nhmrcOutcomesGenderPostDoc}
\alias{nhmrcOutcomesGenderPostDoc}
\title{NHMRC Outcomes Gender PostDoc.}
\format{Data frame with columns
\describe{
\item{year}{Year}
\item{ci}{CI group (CIA, CIB...)}
\item{gender}{Gender}
\item{stage}{Stage (applied, funded, unfunded or profiles) - profiles = number of applicants}
\item{value}{Average years postdoc for stage}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/project_grant_funded_rate_gender_new_investigator_and_teaching_load_140218.xlsx
}
\usage{
nhmrcOutcomesGenderPostDoc
}
\description{
NHMRC project grants new investigators average years postdoc at time of application.
}
\examples{
  nhmrcOutcomesGenderPostDoc \%>\% glimpse()
}
\keyword{datasets}
