% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcOutcomesGenderRatesBS.R
\docType{data}
\name{nhmrcOutcomesGenderRatesBS}
\alias{nhmrcOutcomesGenderRatesBS}
\title{NHMRC Outcomes Gender Rates BS.}
\format{Data frame with columns
\describe{
\item{year}{Year}
\item{ci}{CI group (CIA, CIB...)}
\item{gender}{Gender}
\item{stage}{Stage (applied, funded)}
\item{value}{Count for each stage}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/project_grant_funded_rate_gender_new_investigator_and_teaching_load_140218.xlsx
}
\usage{
nhmrcOutcomesGenderRatesBS
}
\description{
NHMRC project grants new investigators funding rates basic science.
}
\examples{
  nhmrcOutcomesGenderRatesBS \%>\% glimpse()
}
\keyword{datasets}
