% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcOutcomesGenderScheme.R
\docType{data}
\name{nhmrcOutcomesGenderScheme}
\alias{nhmrcOutcomesGenderScheme}
\title{NHMRC Outcomes Gender Scheme.}
\format{Data frame with columns
\describe{
\item{year}{Year}
\item{scheme}{Scheme}
\item{category}{Category}
\item{gender}{Gender}
\item{stage}{Stage (applied or funded)}
\item{value}{Amount awarded}
}}
\source{
2012 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/wihs_table_5_140317v2.pdf

2013 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/wihs_table_1_140317v2.pdf

2014 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/table_1_funding_outcomes_by_gender_scheme_141205.pdf

2015 https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/webchange_wihs_2015_outcomes_by_gender_table_1.pdf
}
\usage{
nhmrcOutcomesGenderScheme
}
\description{
NHMRC funding outcomes by gender for each scheme.
}
\examples{
  nhmrcOutcomesGenderScheme
}
\keyword{datasets}
