% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcOutcomesGenderTeaching.R
\docType{data}
\name{nhmrcOutcomesGenderTeaching}
\alias{nhmrcOutcomesGenderTeaching}
\title{NHMRC Outcomes Gender Teaching.}
\format{Data frame with columns
\describe{
\item{year}{Year}
\item{gender}{Gender}
\item{teaching}{Teaching load}
\item{stage}{Stage (funded, not funded)}
\item{value}{Count for each stage}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/research/women_in_science/project_grant_funded_rate_gender_new_investigator_and_teaching_load_140218.xlsx
}
\usage{
nhmrcOutcomesGenderTeaching
}
\description{
NHMRC project grants CIA Go8 Universities 2010 to 2012 by teaching load
}
\examples{
  nhmrcOutcomesGenderTeaching \%>\% glimpse()
}
\keyword{datasets}
