% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcGrantSubtype.R
\docType{data}
\name{nhmrcGrantSubtype}
\alias{nhmrcGrantSubtype}
\title{NHMRC Expenditure by Grant Subtype.}
\format{Data frame with columns
\describe{
\item{year}{Calendar year}
\item{main_funding_group}{Type of funding}
\item{lower_grant_type}{Grant type}
\item{grant_sub_type}{Grant subtype}
\item{value}{Expenditure for grant subtype}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/dataset/2015/2._summary_tables_2000_-_2015.xlsx
}
\usage{
nhmrcGrantSubtype
}
\description{
Summary of expenditure by grant subtype and year.
}
\examples{
  nhmrcGrantSubtype \%>\% glimpse()
}
\keyword{datasets}
