% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcRegion.R
\docType{data}
\name{nhmrcRegion}
\alias{nhmrcRegion}
\title{NHMRC Expenditure by Region.}
\format{Data frame with columns
\describe{
\item{calendar_year}{Calendar year}
\item{state}{State or territory}
\item{value}{Expenditure for state or territory}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/dataset/2016/all_grants_2000-2016.xlsx
}
\usage{
nhmrcRegion
}
\description{
Summary of expenditure by state/territory and year.
}
\examples{
  nhmrcRegion
}
\keyword{datasets}
