% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmrcSector.R
\docType{data}
\name{nhmrcSector}
\alias{nhmrcSector}
\title{NHMRC Expenditure by Sector.}
\format{Data frame with columns
\describe{
\item{calendar_year}{Calendar year}
\item{sector}{Sector}
\item{value}{Expenditure for sector}
}}
\source{
https://www.nhmrc.gov.au/_files_nhmrc/file/grants/dataset/2015/2._summary_tables_2000_-_2015.xlsx
}
\usage{
nhmrcSector
}
\description{
Summary of expenditure by sector and year.
}
\examples{
  nhmrcSector \%>\% glimpse()
}
\keyword{datasets}
